Fixing test-common.py
