+++ kubeflow/spark/test/common.py
@@ -17,10 +17,6 @@
 from dataclasses import dataclass, field
 from typing import Any
 
-from kubeflow_spark_api import models
-
-from kubeflow.spark.backends.kubernetes import constants
-
 # Common status constants
 SUCCESS = "success"
 FAILED = "failed"
@@ -44,57 +40,3 @@ class TestCase:
     expected_output: Any | None = None
     expected_error: type[Exception] | None = None
     __test__ = False
-
-
-def get_spark_application(
-    name: str,
-    namespace: str = DEFAULT_NAMESPACE,
-    state: str | None = "SUBMITTED",
-) -> models.SparkV1beta2SparkApplication:
-    """Create a mock SparkApplication model for testing."""
-    return models.SparkV1beta2SparkApplication(
-        api_version=f"{constants.SPARK_APPLICATION_GROUP}/{constants.SPARK_APPLICATION_VERSION}",
-        kind=constants.SPARK_APPLICATION_KIND,
-        metadata=models.IoK8sApimachineryPkgApisMetaV1ObjectMeta(
-            name=name,
-            namespace=namespace,
-        ),
-        spec=models.SparkV1beta2SparkApplicationSpec(
-            type="Python",
-            mode="cluster",
-            spark_version="4.0.1",
-            image="spark:latest",
-            main_application_file="s3://job.py",
-            driver=models.SparkV1beta2DriverSpec(
-                cores=1,
-                memory="1g",
-            ),
-            executor=models.SparkV1beta2ExecutorSpec(
-                cores=1,
-                memory="1g",
-                instances=1,
-            ),
-        ),
-        status=(
-            models.SparkV1beta2SparkApplicationStatus(
-                application_state=models.SparkV1beta2ApplicationState(
-                    state=state,
-                ),
-                driver_info=models.SparkV1beta2DriverInfo(
-                    pod_name=f"{name}-driver",
-                ),
-            )
-            if state is not None
-            else None
-        ),
-    )
-
-
-def get_spark_application_list(
-    items: list[models.SparkV1beta2SparkApplication],
-) -> models.SparkV1beta2SparkApplicationList:
-    """Create a SparkApplicationList for testing."""
-
-    return models.SparkV1beta2SparkApplicationList(
-        items=items,
-    )
